feat: normalize image type and file extension
Certain search engines do not sniff the image type and will show
an error if the filename has no extension, or if the image is uploaded
with an incorrect content type.

This change checks the magic number of the image and adjusts
the blob type and file extension before uploading.